feat(batchmail): refactor batchmail components
