+++ frontend/src/components/batchmail/types.ts
@@ -1,0 +1,8 @@
+export type RenderedEmail = {
+  to: string;
+  name?: string;
+  subject?: string;
+  html: string;
+};
+
+export type RecipientScope = "all" | "registered" | "pending";
